requirements.txt
openreview-py>=1.15.0
fpdf2>=2.7.9
playwright>=1.44.0
pandas>=2.0.0
scipy>=1.11.0
matplotlib>=3.8.0
seaborn>=0.13.0
openai>=1.30.0
python-dotenv>=1.0.0
tqdm>=4.66.0
